[Update]: Fix Meeting Service. Remove unused directives. [ArtsiomRusak]

+++ DumaProject/DumaProject.BLInterfaces/BLLInterfaces/IMeetingService.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using DumaProject.Core.Entities;
 
 namespace DumaProject.BLInterfaces.BLLInterfaces
@@ -11,11 +8,11 @@ namespace DumaProject.BLInterfaces.BLLInterfaces
     {
         Meeting SetMeetingForCommission(DateTime date, string venue, int durationInMinutes, Commission commission);
         Meeting SetMeetingForCommission(Meeting meeting, Commission commission);
-        Meeting GetMeetingById(int meetingId);
-        void AddParticipantsForMeeting(int meetingId, ICollection<Member> participants);
-        void ExcludeParticipants(int meetingId, ICollection<Member> participants);
+        Meeting GetMeetingById(long meetingId);
+        void AddParticipantsForMeeting(long meetingId, ICollection<Member> participants);
+        void ExcludeParticipants(long meetingId, ICollection<Member> participants);
         HashSet<Meeting> GetMeetingsForCommission(Commission commission);
-        void CancelMeeting(int meetingId);
-        void EndMeeting(int meetingId);
+        void CancelMeeting(long meetingId);
+        void EndMeeting(long meetingId);
     }
 }
